Add files via upload
This project is a simple Python script that can be used to sniff network packets. The goals of the project are to provide a simple way to sniff packets and to learn more about how network packets work. The challenges I faced were learning how to use the Python socket library and how to parse network packet headers. I overcame these challenges by reading the Python documentation and by searching online for help. I learned a lot about how network packets work by working on this project. I also learned how to use the Python socket library. In the future, I plan to add more features to this project, such as the ability to filter packets and the ability to save packets to a file.